make: initial Rust language support
The idea is to integrate rust into the existing build system as well as
possible and to not use a separate buildsystem like cargo or xargo.

Makefile.base has been modified to include support for main.rs and
lib.rs files and a Makefile.rust has been added which sets a few rust
related variables. Besides target files for the cortex family have been
added this are used and parsed by the rust compiler.

In addition to that a rust/ subdirectory has been added to pkg/ and
contains a makefile for easily porting existing cargo crates to RIOTs
makefile based pkg system.