pkg/rust/core: new package
This is a package for the core crate which resemble the rust core
library. The rust core library is the dependency-free foundation of the
rust standard library.

It is needed because we built riot rust application we the
`#![no_std]`-Feature because we can't really use the rust standard
library itself since it is too fat for our little MCUs.