cpu/stm32f1: Add rust/periph crate
This is needed to export various values defined in
`cpu/stm32f1/include/periph_cpu.h` to Rust. Of cause this introduces
code duplicates since the values are now defined in two files.

Maybe we should consider automatically generating this rust crate with a
tool from the C header file?