Add primitive support for the native cpu
This cpu doesn't differ at all from rust/basecpu. We could add a type
alias for each type from rust/basecpu to this crate but this doesn't
seem to make any sense therefore we decided to use symlinks instead
which is absolutely ugly and stupid but good enough for now.