Move all rust wrappers to a rust/ subdirectory
Also add a RUSTRIOT environment variable which contains an absolute path
to this directory. This was a suggestion made by @kaspar030 implementing
it is useful because we don't pollute the other directories.

Besides it makes it easier to track the changes that we make on existing
files. The only rust subdirectory I didn't move is in pkg/ since the
USEPKG system is not as modular as the USEMODULE one.

Fixes #27