Basic wrappers for gcoap
This only includes wrappers for writing response handlers. Wrappers for
functions needed to write a CoAP client are not included since this is
at the moment not useful to us and we are planning to replace this
wrapper with a custom CoAP implementation in rust in the long run
anyhow.